+++ packages/server/src/index.ts
@@ -1,3 +1,3 @@
-export { createServer } from "./server/createServer";
-export { NonameServer } from "./server/NonameServer";
+export { createServer } from "./server/create-server";
+export { NonameServer } from "./server/noname-server";
 export type { ServerInstance, ServerLogEvent, ServerLogger, ServerOptions } from "./types";
