+++ packages/server/src/server/create-server.ts
@@ -1,5 +1,5 @@
 import type { ServerInstance, ServerOptions } from "../types";
-import { NonameServer } from "./NonameServer";
+import { NonameServer } from "./noname-server";
 
 export function createServer(options: ServerOptions = {}): ServerInstance {
 	return new NonameServer(options);
