+++ packages/server/src/services/room-service.ts
@@ -1,5 +1,5 @@
 import type { Client, Room } from "../types";
-import { ServerState } from "../state/ServerState";
+import { ServerState } from "../state/server-state";
 import { sendMessage } from "../utils/send";
 
 export class RoomService {
